fix(creusot): remove unsupported f64-to-f32 casts in ui_types proofs
Creusot does not support FloatToFloat pointer casts (as f32). The two
bounds_to_size proofs used (f64).abs() as f32 which blocked compilation
of all 31 Creusot modules. Removed the cast — the abs() result stays
as f64 which is sufficient for the non-negativity proof.

+++ crates/elicitation_creusot/src/ui_types.rs
@@ -277,7 +277,7 @@ pub fn verify_render_stats_eq() -> bool {
 pub fn verify_bounds_width_non_negative() -> bool {
     let x0: f64 = 10.0;
     let x1: f64 = 110.0;
-    let w = (x1 - x0).abs() as f32;
+    let w = (x1 - x0).abs();
     w >= 0.0
 }
 
@@ -290,7 +290,7 @@ pub fn verify_bounds_width_non_negative() -> bool {
 pub fn verify_bounds_reversed_non_negative() -> bool {
     let x0: f64 = 200.0;
     let x1: f64 = 50.0;
-    let w = (x1 - x0).abs() as f32;
+    let w = (x1 - x0).abs();
     w >= 0.0
 }
